Modified code to recover old features.
Now you can query any field and this must not be textual.
Till now, any field was appended by "_t" as it was thought to be textual.

This was not this way on the original implementation and it is useful if you want to filter
the search by other things but text.

I found that "bug" querying time ranges.